MultipleDice 클래스 삭제, Dices 클래스로 통일
사유는 함수 호출의 편리성을 위해서
기존 Dice -> MultipleDice -> Player 형식은 비효율적
